backend/pyproject.toml
[project]
name = "ai-agent-backend"
version = "0.1.0"
description = "Backend service for the AI Agent project — task coordination and LLM orchestration"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "requests>=2.32.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.25",
  "aiosqlite>=0.20.0",
  "chromadb>=0.5.5",
  "faiss-cpu>=1.8.0",
  "redis>=5.0.3",
  "rq>=1.16.2",
  "pytesseract>=0.3.10",
  "Pillow>=10.2.0",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.5",
  "black>=24.3.0",
  "flake8>=7.0.0",
  "mypy>=1.9.0",
  "tenacity>=8.3.0",
  "python-multipart>=0.0.20",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
ignore_missing_imports = true
strict = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]
